% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello world
This is a function to say hello to the world.}
\usage{
hello()
}
\value{
hello returns a string named hello world.
}
\description{
Hello world
This is a function to say hello to the world.
}
\examples{
library(learning)
hello()
}
